script: Use sh instead of execline
execline aur package is not very stable. It will often stop building for
no reason. Even if I like execline more than sh, sh is sure to be
around, while execline is not. If the maintainer stops working on it, it
is not popular enough to be maintained by someone else.

Goodbye execline. It was fun